Replaced the seeding of test with Factory_bot
