+++ app/models/verb_noun_link.rb
@@ -1,5 +1,4 @@
 class VerbNounLink < ApplicationRecord
   belongs_to :noun
   belongs_to :verb
-  validates :noun_id, :verb_id, presence: true
 end
