+++ db/migrate/20191112100304_add_is_a_to_noun.rb
@@ -1,5 +1,5 @@
 class AddIsAToNoun < ActiveRecord::Migration[5.2]
   def change
-    add_column :nouns, :is_a, :string
+    add_column :nouns, :kind, :string
   end
 end
