+++ spec/models/verb_noun_link_spec.rb
@@ -5,9 +5,4 @@ RSpec.describe VerbNounLink, type: :model do
     it { should belong_to(:verb) }
     it { should belong_to(:noun) }
   end
-
-  describe 'validations' do
-    it { should validate_presence_of(:verb_id)}
-    it { should validate_presence_of(:noun_id)}
-  end
 end
